feat(phone): AVO-068 Phase 1 - extract diff widgets and add expandable commit tiles
- Extract DiffView, DiffEntryTile, BinaryFileIndicator, DiffHunkView, DiffLineView
  into new shared widgets/diff_widgets.dart
- Update CommitDiffScreen to use extracted public widgets
- Add expandable commit tiles with lazy-loaded inline diffs in BranchDetailScreen
- Apply same expandable pattern to CommitHistoryScreen
- Diff is fetched on-demand when tile is expanded (per-item, not pre-fetched)
- Per-item expansion state allows multiple commits expanded simultaneously
